Added deleting of entities
