+++ src/repositories/WalletRepository.kt
@@ -1,16 +1,18 @@
 package ru.banking.repositories
 
 import database.DatabaseFactory
+import database.DatabaseFactory.dbQuery
 import org.jetbrains.exposed.sql.*
+import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
 import ru.banking.database.Wallet
 import ru.banking.database.Wallets
 
 class WalletRepository {
-    suspend fun getAllWallets(): List<Wallet> = DatabaseFactory.dbQuery {
+    suspend fun getAllWallets(): List<Wallet> = dbQuery {
         Wallets.selectAll().map { toWallet(it) }
     }
 
-    suspend fun getWallet(id: Long): Wallet? = DatabaseFactory.dbQuery {
+    suspend fun getWallet(id: Long): Wallet? = dbQuery {
         Wallets.select {
             (Wallets.id eq id)
         }.mapNotNull { toWallet(it) }
@@ -19,7 +21,7 @@ class WalletRepository {
 
     suspend fun addWallet(wallet: Wallet): Wallet {
         var key = 0L
-        DatabaseFactory.dbQuery {
+        dbQuery {
             key = (Wallets.insert {
                 it[currency] = wallet.currency
                 it[ballance] = wallet.ballance
@@ -30,12 +32,18 @@ class WalletRepository {
     }
 
     suspend fun addWallets(wallets: List<Wallet>) {
-        DatabaseFactory.dbQuery {
+        dbQuery {
             Wallets.batchInsert(wallets) { wallet ->
                 this[Wallets.currency] = wallet.currency
                 this[Wallets.ballance] = wallet.ballance
                 this[Wallets.customerId] = wallet.customerId
             }
+        }
+    }
+
+    suspend fun deleteWallet(id: Long): Boolean {
+        return dbQuery {
+            Wallets.deleteWhere { Wallets.id eq id } > 0
         }
     }
