+++ src/services/CustomerService.kt
@@ -1,5 +1,10 @@
 package ru.banking.services
 
+import database.DatabaseFactory
+import javafx.application.Application.launch
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.async
+import kotlinx.coroutines.runBlocking
 import org.jetbrains.exposed.dao.EntityID
 import ru.banking.database.Citizenship
 import ru.banking.database.Customer
@@ -22,15 +27,33 @@ class CustomerService(
         return customerRepository.getCustomer(customerId)
     }
 
-    suspend fun addWalletsToCustomer(wallets: List<Wallet>) {
-        walletRepository.addWallets(wallets)
-    }
-
     suspend fun getAllCustomers(): List<Customer> = customerRepository.getAllCustomers()
 
     suspend fun getCustomerWithWallets(customerId: Long): CustomerDTO? {
         val customers = customerRepository.getCustomerWithWallets(customerId)
-        val firstCustomer = customers?.get(0)!!
-        return CustomerDTO(firstCustomer, customers.mapNotNull { it.wallet }.map { WalletDTO(it) })
+        return if (customers?.size == 0) {
+            null
+        } else {
+            val firstCustomer = customers?.get(0)!!
+            CustomerDTO(firstCustomer, customers.mapNotNull { it.wallet }.map { WalletDTO(it) })
+        }
+    }
+
+    suspend fun removeCustomer(customerId: Long): Boolean {
+        return DatabaseFactory.dbQuery {
+            runBlocking {
+                val customer = getCustomerWithWallets(customerId)
+                customer?.wallets?.forEach { wallet -> wallet.id?.let { walletRepository.deleteWallet(it) } }
+                customerRepository.deleteCustomer(customerId)
+            }
+        }
+    }
+
+    suspend fun addWalletsToCustomer(wallets: List<Wallet>) {
+        walletRepository.addWallets(wallets)
+    }
+
+    suspend fun removeWallet(walletId: Long): Boolean {
+        return walletRepository.deleteWallet(walletId)
     }
 }
